updated the post liking system optimistically using rtk query for giving better user experience
