+++ src/pages/home/componnents/PostCard.jsx
@@ -15,9 +15,7 @@ const PostCard = ({
   likes,
   user = null,
 }) => {
-
   const isLiked = likes?.indexOf(user.email);
-
   const getPostAge = moment(postTime).fromNow();
 
   return (
