+++ src/pages/home/home/Home.jsx
@@ -38,7 +38,6 @@ const Home = () => {
   }
 
   if (!isLoading && isSuccess) {
-    console.log(posts, "posts_______________________________________________");
     content = posts.map((post) => (
       <PostCard
         key={post._id}
