+++ src/redux/features/post/postApi.js
@@ -1,17 +1,12 @@
 import { apiSlice } from "../api/apiSlice";
 
-export const authApi = apiSlice.injectEndpoints({
+export const postApi = apiSlice.injectEndpoints({
   endpoints: (builder) => ({
     getPosts: builder.query({
       query: () => ({
         url: "/posts/get-followings-posts",
-
         method: "GET",
       }),
-      providesTags: (result) =>
-        result
-          ? result.map((post) => ({ type: "Posts", id: post._id }))
-          : ["Posts"],
     }),
     createPost: builder.mutation({
       query: ({ newPost }) => ({
@@ -29,10 +24,32 @@ export const authApi = apiSlice.injectEndpoints({
         url: `/posts/like/${postId}`,
         method: "PATCH",
       }),
-      invalidatesTags: (data) => [{ type: "Posts", id: data._id }],
+      async onQueryStarted({ postId }, { getState, dispatch, queryFulfilled }) {
+        const userEmail = getState()?.auth?.user?.email;
+
+        // optimistic cache update start
+        const patchResult = dispatch(
+          postApi.util.updateQueryData("getPosts", undefined, (draft) => {
+            const draftPost = draft.find((p) => p._id === postId);
+            const isLikeExist = draftPost.likes.indexOf(userEmail);
+            if (isLikeExist !== -1) {
+              draftPost.likes.splice(isLikeExist, 1);
+            } else {
+              draftPost.likes.push(userEmail);
+            }
+          })
+        );
+        // optimistic cache update end
+
+        try {
+          await queryFulfilled;
+        } catch (error) {
+          patchResult.undo();
+        }
+      },
     }),
   }),
 });
 
 export const { useGetPostsQuery, useCreatePostMutation, useLikeMutation } =
-  authApi;
+  postApi;
