+++ src/redux/store.js
@@ -1,11 +1,13 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { apiSlice } from "./features/api/apiSlice";
 import authSlice from "./features/auth/authSlice";
+import userSlice from "./features/user/userSlice";
 
 export const store = configureStore({
   reducer: {
     [apiSlice.reducerPath]: apiSlice.reducer,
     auth: authSlice,
+    user: userSlice,
   },
   devTools: import.meta.env.NODE_ENV !== "production",
   middleware: (getDefaultMiddlewares) =>
