+++ src/routes/Routes.jsx
@@ -12,6 +12,7 @@ import PrivateRoute from "./PrivateRoute";
 import PublicRoute from "./PublicRoute";
 import Profile2 from "../pages/profile/Profile2";
 import Info from "../pages/info/Info";
+import InfoRouteHandle from "./InfoRouteHandle";
 
 const Routes = createBrowserRouter([
   {
@@ -46,7 +47,11 @@ const Routes = createBrowserRouter([
       },
       {
         path: "info",
-        element: <Info />,
+        element: (
+          <InfoRouteHandle>
+            <Info />
+          </InfoRouteHandle>
+        ),
       },
     ],
   },
